requirements.txt
django-environ
django
djangorestframework
djangorestframework_simplejwt
